Add a new function for gathering entries that is not only compatible with the theme, but MU plugins, and those that may be nested MU plugins.
Make sure old file is backwards compatible
